requirements.txt
flask
pytube
openai-whisper
googletrans==4.0.0-rc1